requirements.txt
# Core FastAPI
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.29.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Auth & Security
python-multipart>=0.0.6,<0.1.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# HTTP Client
httpx>=0.25.0,<0.29.0

# Supabase
supabase>=2.5.0,<3.0.0

# Config
python-dotenv>=1.0.0,<2.0.0

# AI SDKs
openai>=1.12.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0

# WebSockets (Optimized for Supabase and Realtime compatibility)
websockets>=13.0,<16.0

# Snowflake
snowflake>=1.11.0,<2.0.0
snowflake-connector-python>=3.10.0,<4.0.0
snowflake-snowpark-python>=1.18.0,<2.0.0
snowflake-connector-python[pandas]>=3.10.0,<4.0.0

# Pydantic
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Crypto stack
cryptography>=41.0.0,<43.0.0
charset-normalizer>=2.0.0,<4.0.0
